Local training policy: ensure scheduled queue size is bounded
When we have unfinished tasks that we don't need to wait for, we risk entering into a timing issue where the rate of queueing more work is larger than the rate of clearing worker processes.

To both keep the system well-performing and robust to such corner cases, we track the unfinished work. We require the data collection plugin (e.g. inliner, regalloc) to have a timeout - so plugins are guaranteed to finish, once a worker acquires them. If a watermark of unfinished work is hit, we report and wait for it to drain (up to that watermark). If this happens frequently, the user should increase deadlines or check the corpus (maybe there are too many very large modules that take a long time to compile)

This design quite easily lends us the opportunity to use the data collected from work from previous training ('stale' data). The option is available via a flag.

Separately, when finishing all work, we also need to graciously close and join the current pool, otherwise we get errors due to timing between: pool workers exiting, main process exiting, and pool workers' communication channels being severed because of the latter. Note that in google3, forcefully terminating the pool causes the spanned processes (clang) to crash (expected), but that also appears to pull down the main process - we should investigate this separately - it'd enable us exit more quickly.