Decouple our tooling from problem solvers (generalize 'problem type')
We currently check for known problem types (e.g. 'inlining' or 'regalloc') and select the appropriate CompilationRunner implementation and ML configuration.

This won't scale moving forward, and this patch addresses this via dependency injection (gin).

The abstraction being injected is an instance of ProblemConfiguration. This instance can be queried for a CompilationRunner instance, and for the ML configuration (preprocessing layer creator and signature spec). All parameters for these (e.g. clang_path for the CompilationRunner) are also gin-injected.

A second reason for having modules as implementations of ProblemConfiguration is the fact that we typically need quite a few artifacts to fully define such an implementation: an implementation of a CompilationRunner; ML configuration (the config.py); and gin settings. The wrapping module forms, thus, an intuitive component model, and forms a single extension point (all a tool needs configured is 1 module name)

A tool just needs to do this:

in python: problem_config = config.ProblemConfiguration() ... and then get what it needs from problem_config

take --gin_bindings flags. The user just needs to:

--gin_bindings=configs.ProblemConfiguration.module_name="'the.module.name'"

e.g.:

--gin_bindings=configs.ProblemConfiguration.module_name="'google3.research.sir.compiler_opt.rl.inlining'"

This refactoring also opens the opportunity to avoid copying around common abseil flags, e.g. for clang path; instead, they can be gin-configured. In addition, this simplifies special cases, like inlining needing llvm-size (inlining for speed will likely need llvm-objdump... etc): instead of needing to add the union of all flags to each tool, we rely on gin and define a gin macro for each, thus allowing us to replace --clang_path= with --gin_bindings="'clang_path='" (for example). This trades off a bit of verbosity in cmd line flags for the aforementioned advantages.