Fixed regalloc behavioral cloning
Before this patch, the input size of the model created
using behavioral cloning would not be compatible with
the local trainer(configured with ppo_nn_agent.gin) due
to the with_sqrt flag not explicitly being set to false
in the regalloc behavioral cloning config. This patch
adjusts the behavioral_cloning_nn_agent.gin config to
more precisley match the ppo_nn_agent.gin config in this
area so that the models exported by one are compatible
with the other.