Prevent TF from logging info messages during data collection
Currently, whenever a TF model gets loaded in by LLVM in the
development mode advisors, TF outputs a slew of info messages
to STDOUT that aren't particularly useful. This patch gets
rid of those messages by setting an environment variable when
running the compilation commands that only enables warnings
and errors to be printed to the console. This makes it much
easier to see progress/results in several of the scripts within
this repository such as train_locally.py and
generate_default_trace.py when grabbing performance data for a
policy.