Pass serialized policies to workers
This sets up the trainer to work with potentially remote workers by
passing a serialized object to them. Each worker, then, serializes the
policy locally, after which data collection continues as usual.

The main inefficiency is that we save the policy per request handler,
and even if the training is local. Policies are small objects, however
(at least currently), so this shouldn't be cause of concern. Upon
measuring this for 100 and then 500 modules, with 128 workers:

- 100 modules: ~7% regression in the compile time component of training
- 500 modules: ~9% improvement

At this point, I'd take neither regression nor improvement as
meaningful, and since, additionally, compilation time is about 1/10 of
overall time, the complexity tradeoff of a design optimizing for the
local case is probably not worth it.