Explicitly model 'replace' flags (#139)
* Explicitly model 'replace' flags

Using 'delete' and 'add' for flags that must actually be present - such
as -fthinlto-index, or, later, profiles - is a bug farm. This patch
explicitly models such flags as something that cannot be 'deleted' or
'added', and must be somehow converted (contextualized to the location
of the files(presumably) used as values of those flags). The conversion
rule is a formatting string which uses a 'context' value, which will
have fields populate at the time of the conversion.

Some tests had to have the order of some flags adjusted a bit as a
NFC side-effect.

* Rebase and fix