Removed num_modules flag from RL training in demo
We refactored this flag a while ago to be specified in the gin config
for the relevant problem to improve reproducibility, especially in the
regalloc case, but I forgot to remove the flag from the demo. This patch
removes the flag from the demo, making everything (hopefully) plug and
play.