Removed tensorflow_c_lib_path flag from LLVM build for benchmarking (#150)
Should help with the other half of #129. The benchmarking code currently
sets the TENSORFLOW_C_LIB_PATH CMake variable when building LLVM.
However, this doesn't make sense going forward as we have shifted mostly
everything over to TFLite and we don't need to have MLGO development
mode enabled for benchmarking in either case. This patch explicitly
removes the flag so that the benchmarks simply build LLVM in MLGO
release mode.