Introduce a Nesting Worker
In some distributed object middleware, there's a limitation that one
remote process host one object instance only. During training, we have
'lull' periods when we don't collect data for an experiment (because we
run the training algo on the data). If we run a bunch of related
experiments, we can reuse the remote resource during that time. This can
be advantageous if the total resource utilization of 2N remote tasks used
at (say) 50% is higher than that of N tasks used at 100% (which is
usually the case - there's some book keeping cost)

The design uses a special kind of worker which manages a bunch of
underlying workers to which it delegates all calls.

The current patch just introduces the bare minimum support. Subsequent
patches will add better error handing (e.g. worker id doesn't exist),
and potentially some load balancing support, if needed.