Fix Python Versioning on CI (#194)
My commit to move everything to pipenv also updated the CI to use
pipenv, but pipenv is picking up the wrong Python version when creating
the virtual environment, so all of the jobs are using Python 3.10. This
patch fixes that behavior by having pipenv install the packages to the
system rather than creating a virtual environment.