Remove protobuf build from build_tflite.sh script
Now that the transition to the new dependency free logger is
complete, there's no longer a need to have a protobuf build
in order to build LLVM in MLGO development mode. This patch
removes the protobuf build from build_tflite.sh which will
remove the extra protobuf build from both the buildbots and
the development mode Docker image.