Fix bug with positional arguments in the LocalWorkerPool.
Positional arguments for worker_class wouldn't work at all when creating
a local worker pool, which surprisingly didn't cause any issues
previously because we only ever instantiated the pool with kwargs for
the worker_class. This is because of how the function signatures for
_run and _run_impl were written in local_worker_manager.py. The modified
test in this commit would previously fail without the changes in
local_worker_manager.py.