Rebase Chromium Patches
There have been some changes within the Chromium build system that have
broken these patches, making work on Chromium with MLGO (especially
through the demo) more difficult than it needs to be.

This patch rebases these patches so they apply cleanly (when performed
in order) against a current (as of this commit) checkout of the Chromium
source code.