Fix extract_ir.py for macos.
For whatever reason, the section names for .llvmcmd and .llvmbc are different
on macos. One problem is that macos uses macho object files instead of elf, so
in addition to the section names there are segment names that need to be
specified. The other problem is that regardless of the segments, the section
names are different. This patch makes the segment name configurable by a flag,
and gives sensible defaults for linux + documentation for macos.