Fix caching in CI (#262)
This fixes the caching for python dependencies in the CI by making
pipenv install in a virtual environment and telling the setup-python
action to cache this venv as it doesn't currently support the global
pipenv cache. This increases the speed of some of the fast jobs (like
the yapf format checker) quite a bit. Formatting went from about 2.5min
to about 0.7min in my testing.