Add directory input to extract_ir.py
This patch adds in a directory input to extract_ir.py where the script
does a glob for *.o files inside the specified directory. This is useful
in cases where it is easy to specify specific flags through a build
system (like the necessary -Xclang -fembed-bitcode=all), but the build
system produces no manifest of compilation commands like CMake's
compile_commands.json. This isn't reccomended to use in place of a
proper manifest but is necessary for supporting build systems like Make.