Add docstrings to extract_ir_lib.py
Now that extract_ir_lib.py is being used in other projects through the
Pip package rather than just the ml-compiler-opt CLI tooling, it would
be good to have better documentation on how to use these functions. This
commit adds in doc strings for some of the functions that are most
likely to be used in downstream projects and fixes up some other
docstrings.

This was motivated by not passing an empty string instead of None to
cmd_filter which led to some annoying debugging when other issues were
at play.