Move extract_artifacts to inside run_extractions
This is a minor refactoring that simplifies the code base for future
work. It doesn't make a lot of sense to keep this as a separate function
and then use a functools.partial to setup the correct command line
arguments when we can just do it explicitly using a subfunction.