Fix run_extraction after earlier patch
My refactoring earlier in 32064da4250c37850f376663eff1cf1b79ab228a ended
up breaking the extract_ir.py script as local objects can't be pickled.
This patch changes everything back to using a functools.partial() but
maintains the spirit of the original refactoring by still eliminating
the global helper function.

I would like to add in a test case for this but it would have to be an
integration style test to get the setup correct and we don't have the
infrastructure to do that currently.