Minor fixups in extract_ir.py
This patch includes some (very) minor formatting changes. The flags are
all grouped together, the single required flag statement is changed to
mark_flag_as_required rather than passing an array with a single element
to the flags version. This also fixes a typo that I made in patch
earlier this week.