Limit Github Actions packaging job to only run on main repo
Currently, the github action that does a scheduled publish of the
package tries to run on all forks (or at least mine) when it is present
in the main branch. This isn't desirable as even though nothing will
actually get pushed to PyPI as the repository secrets aren't present,
even just trying it will cause a notification for a failed job to be
sent at 00:00 UTC every day. This patch fixes that behavior by
restricting the twine upload step to only run on the main branch in the
google/ml-compiler-opt repo (i.e., no forks) as Github doesn't seem to
have better controls to enforce this sort of behavior.