Clean apt packages/cache after build
This commit adds in a step at the end of the development Dockerfile to
clean the apt repositories/cache/unused packages from the image to make
it a little bit smaller and more production like. This isn't super
necessary for a development image that likely won't leave the machine
it's built on, but this is best practice and is barely noticeable in
terms of image build time.