Make combine_corpus only look at folders with a corpus description
Currently the combine corpus tooling just looks through all files and
folders in a directory, assuming them all to be folders with a
corpus_description.json in them. We don't actually throw an error (only
logging one), but it makes sense to be more robust and not log an error
when we do something like iterate over a logging file. This patch also
removes the tensorflow file code from this library which provides a
couple benefits. It means we no longer have to load tensorflow to call
the tool which is a significant time-savings and developer tooling like
pylint works better on the standard library functions.