Make combine_corpus only look at folders with a corpus description (#290)
Currently the combine corpus tooling just looks through all files and
folders in a directory, assuming them all to be folders with a
corpus_description.json in them. We don't actually throw an error (only
logging one), but it makes sense to be more robust and not log an error
when we do something like iterate over a logging file.