Install pytype on buildbot
This patch installs the development dependencies on the buildbot in
addition to the normal python dependencies so that we have pytype
available. This is necessary for running the pytype tests recently added
on the LLVM side so we can prevent regressions while waiting for
precommit CI to be implemented on the LLVM side.