Better showcase package installation command
This patch updates the README to better showcase the package
installation commands. Most people probably just want to install
everything and it probably won't hurt if they do so. So make that the
default and then add a clarifying explanation and what is what rather
than telling them to install more packages below.