Restrict push CI to main, fix permissions
This patch makes two primary changes to the CI config. It restricts the
CI running on push events to only run on the main branch. This prevents
the CI from running twice on PRs from branches in the main repository,
which are needed for stacked PRs.

Additionally, this PR sets the permissions at the top of the file, which
is a general best practice for Github workflows for security reasons.