Serialize entire policy in blackbox_learner
This patch adjusts blackbox_learner so that it returns an entire policy
rather than just the bytes of the policy. When actually running
evaluations, we need to writ out the full policy, including the output
spec, to disk so the compiler can pick it up. Before this patch, we were
not passing along the output spec to the worker.