Handle `gin.bind_parameter` in tests
If a test needs to `gin.bind_parameter` it should also clear that, or
make the binding within a `gin.config_scope`, otherwise the bindings may
leak to other tests if all the tests are run by e.g. pytest.

This patch uses both approaches: for tests that set up bindings in `setUp`
we clear in `tearDown`, else we use the scoped approach.

Issue #394