Add a test for trainer metrics
This patch adds a test for the metrics in the trainer class to ensure
that they are actually set. This patch just directly inspects the state
rather than trying to ensure that we log to tensorboard too, but this
should be good enough, and is definitely better than what we had before
(nothing).