Bump python deps
This patch bumps all the python deps to more recent versions. Tensorflow
is not bumped to the latest version as the last tf_agents release is
only compatible with v2.15.0. We need to figure out what to do here, but
this should still be an improvement that we can build off of.

The lockfile here was regenerated with Python 3.10, so probably will not
be compatible with anything older.