[BC] ```keep_temps``` update (#411)
* Rename ```base_path``` to ```persistent_objects_path``` and ```keep_temps``` to ```explicit_temps_dir```.
Makes sure that if ```persistent_objects_path``` is set then ```explicit_temps_dir``` is also set
or neither of them is set. Further, passes ```explicit_temps_dir``` to
```compilation_runner.get_workdir_context``` and ensures that the flag
and argument are not set at the same time.

* Addressing @mtrofin and @boomanaiden154 comments.

* Set explicit_temps_dir to persistent_objects_path+'/temp_dirs' whenever
explicit_temps_dir is not set and persistent_objects_path is set.