Refactor gen_test_model to separate module
This patch factors the _gen_test_model function in policy_saver_test
into a separate module in a new testing subdirectory to facilitate reuse
across tests. This is intended to be used in the regalloc_trace_worker
test which has to create a test model to test that everything works with
TFLite.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/414