Ignore .pytype directory for pylint (#417)
After upgrading pytype, it now writes a bunch of .pyi files into the
.pytype directory. yapf and pylint pick these directories up and report
a ton of issues that are completely meaningless. This patch makes it so
that pylint ignores any files within .pytype. yapf does not have any
option in the config to ignore files, so just blanket running yapf . -ir
or specifying individual files is still necessary, but I think that
should be less interruptive. This is purely a developer workflow
improvement.