log_reader: tighten parsing
The log format has a few places where we insert `\n` for human readability.
They should be checked they are indeed just that character. This will allow us,
in a subsequent PR, to avoid infinitely-blocking cases due to the process
producing the log exiting unexpectedtly, by inserting errors into the reader
to unblock and exit it.

Similarly, checking that the tensor data received matches in size what
was expected.

Refactored a bit the test utility for constructing examples.

commit-id:b1513456