Add initial gin configs for regalloc_trace in ES
This patch adds a set of gin configs that can be used for regalloc_trace ES
training. This are intended mostly to be starting points so that we have
a set of options that work upstream as an example and they can be modified
within experiments as needed. They need to be modified before use, such as
by setting binary and basic block trace paths.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/435