Add convert_to_tflite to policy_utils
This patch takes the function that converts policies serialized to numpy
arrays serialized to bytes into TFLite models to policy_utils. This is
necessary so that we can perform this on the worker rather than trying
to perform it serially on the main invocation.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/447