Move TFLite Policy Generation to ES Workers
This patch moves TFLite Policy Generation for ES to the workers. This
step is reasonably expensive and doing it serially on the main node
takes quite some time with a reasonable number of perturbations. This
patch moves this step to the workers to parallelize it.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/448