Add base class for WorkerPoolManagers (#470)
This patch introduces a base class that worker pool managers can inherit
from. This enforces the interface, particularly for instantiation, that
we have been having a bit of trouble with recently given some recent
refactorings. I've validated that this patch would have caught the
issues that have already been fixed.