Accept decay rate parameter in MockCompilationRunner
The mock used in generate_default_trace_test needs to accept
moving_average_decay_rate as it is automatically passed in by the
instantiation setup in generate_trace in generate_default_trace.py. The
test still passes if it isn't this way, but doesn't actually "compile" any
modules due to the errors, which sort of defeats the point.