Add option to generate_default_trace to output keys (#477)
This patch adds an option to generate_default_trace to output the keys
associated with the examples. This is primarily intended for use in
extracting functions that contain eviction decisions to improve the
efficiency of the new training workflow for ES regalloc trace modelling.
However, this might also be useful in other circumstances if more
introspection into the data is needed.