Name scalar based on percentile to report (#478)
Before this patch, the scalar name would end up as containing the value
of the percentile rather than the percentile that we are trying to
compute. We then end up with a bunch of one point metrics which clutters
the UI and also means we cannot get any information about the actual
percentiles. This patch fixes that by actually naming the scalar after
the percentile that we are trying to compute.