Make extract_functions create output directory if nonexistent
It is a bit annoying when extract_functions fails because the output
path does not exist. This always ends up being a CalledProcessError that
is difficult to debug as stdout/stderr is not logged. Just create the
path because it is probably what the user expected anyways.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/496