Make compile_corpus BC mode produce an actual output corpus
When using the compile_corpus script to partially compile a corpus
(running middle-end optimizations), we need to make sure that the output
is actually another corpus. This entails making sure the file
suffix/extension is the same as the original corpus and that .cmd files
are copied over. This patch adds that funtionality.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/495