Use unittest framework tempdirs in blackbox_learner_test
blackbox_learner_test was using the tempdir (ie /tmp) rather than
absltest managed temp directories that are cleaned up at the end of the
test. This patch fixes it so that we are using the proper temp
directories.

Without this patch some failures for patches up the stack were not
emerging due to files that were leftover in /tmp.

Reviewers: mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/509