Remove act support to simplify github workflow (#512)
This patch removes act support from the main github actions workflow to
make it significantly simpler. This also makes it a little bit faster as
we no longer rely on the envvars job. Act is generally a pain to work
with given the environment only quasi matches that of Github's (hence
all the hacks within this file) and is in general much less convenient
than just running the individual commands within a local terminal
window.