Save models during ES training in a separate thread
This patch makes it so that blackbox_learner saves models during ES
training in a separate thread. There is a bit of additionaly complexity
so that we save the models concurrently while waiting for compilation
results (when the process will be blocked waiting on IO). If we just
spawn jobs whenever we run into performance problems due to the GIL and
are back where we started. We cannot use subprocesses as we cannot
pickle the policy saver function.

This extra complexity is justified through significant performance gains on
training jobs with relatively short step times (up to 20%, still single digit
percentages with step times around 5 minutes).

Reviewers: svkeerthy, mtrofin

Reviewed By: mtrofin

Pull Request: https://github.com/google/ml-compiler-opt/pull/524